IntelliManage/requirement.txt
asyncio
json
os
python-dotenv
openpyxl
fastapi
openai==1.30.5
difflib
logging
httpx